# backend/doc/openapi.yaml
openapi: 3.1.0
info:
  title: Soniq API
  version: 0.1.0
  description: Local-first API for identity, workspaces, recording uploads, processing status, and transcript/summary results.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Same-origin local API server.
security:
  - CookieAuth: []
paths:
...
  /workspaces:
    get:
      summary: List workspaces for the current user
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspaces available to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkspacesResponse'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings:
    get:
      summary: List workspace recordings
      operationId: listRecordings
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
      responses:
        '200':
          description: Recent recordings in the workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRecordingsResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Create recording metadata
      description: Creates a recording session in a workspace without uploading audio or starting processing.
      operationId: createRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordingRequest'
      responses:
        '201':
          description: Recording created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /workspaces/{workspace_id}/recordings/{recording_id}:
    get:
      summary: Get recording
      operationId: getRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/RecordingID'
      responses:
        '200':
          description: Recording found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: soniq_session
  parameters:
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
      description: Workspace identifier.
    RecordingID:
      name: recording_id
      in: path
      required: true
      schema:
        type: string
      description: Recording identifier.
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    WorkspaceRole:
      type: string
      enum: [owner, member]
    Workspace:
      type: object
      required: [id, name, role, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListWorkspacesResponse:
      type: object
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
    WorkflowType:
      type: string
      enum: [meeting, lecture, interview, memo]
    RecordingStatus:
      type: string
      enum: [uploaded, processing, transcribing, summarizing, completed, failed, cancelled]
    Recording:
      type: object
      required: [id, workspace_id, title, status, workflow_type, language, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/RecordingStatus'
        workflow_type:
          $ref: '#/components/schemas/WorkflowType'
        language:
          type: string
        audio_object_key:
          type: string
        audio_content_type:
          type: string
        audio_size_bytes:
          type: integer
          format: int64
        failure_reason:
          type: string
        completed_at:
          type: string
          format: date-time
        failed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRecordingRequest:
      type: object
      required: [workflow_type]
      properties:
        title:
          type: string
        workflow_type:
          $ref: '#/components/schemas/WorkflowType'
        language:
          type: string
          example: en
    ListRecordingsResponse:
      type: object
      required: [recordings]
      properties:
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/Recording'
...
    ErrorResponse:
      type: object
      required: [code, message, status]
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - invalid_credentials
            - user_already_exists
            - validation_failed
            - forbidden
            - not_found
            - method_not_allowed
            - request_too_large
            - unsupported_media_type
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
        message:
          type: string
        status:
          type: integer